#181: Moved initialising `rowCountCache` to `CaffeineAsyncCache`
