+++ app/src/main/scala/org/alephium/explorer/cache/CaffeineAsyncCache.scala
@@ -16,18 +16,60 @@
 
 package org.alephium.explorer.cache
 
-import java.util.concurrent.CompletableFuture
+import java.util.concurrent.{CompletableFuture, Executor}
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 import scala.jdk.FutureConverters._
 
-import com.github.benmanes.caffeine.cache.AsyncLoadingCache
+import com.github.benmanes.caffeine.cache.{AsyncCacheLoader, AsyncLoadingCache, Caffeine}
+import slick.jdbc.PostgresProfile.api._
+
+import org.alephium.explorer.persistence.DBRunner
 
 object CaffeineAsyncCache {
 
   @inline def apply[K, V](cache: AsyncLoadingCache[K, V]): CaffeineAsyncCache[K, V] =
     new CaffeineAsyncCache[K, V](cache)
 
+  /**
+    * Caches the number of rows return by a query.
+    *
+    * @note Slick's [[Query]] type does not implement equals or hashCode so
+    *       this cache will not work as expected on dynamically generated queries.
+    *       The queries to cache should be static i.e. should have the same memory
+    *       address.
+    *
+    *       {{{
+    *          //OK
+    *          cacheRowCount.get(mainChainQuery)
+    *          cacheRowCount.get(mainChainQuery)
+    *
+    *          //OK
+    *          cacheRowCount.get(BlockHeaderSchema.table)
+    *          cacheRowCount.get(BlockHeaderSchema.table)
+    *
+    *          //NOT OK
+    *          cacheRowCount.get(BlockHeaderSchema.table.filter(_.mainChain))
+    *          cacheRowCount.get(BlockHeaderSchema.table.filter(_.mainChain))
+    *       }}}
+    *
+    * @param runner  Allows executing the input [[slick.lifted.Query]].
+    * @param builder Configured [[Caffeine]]'s cache instance.
+    */
+  @SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
+  def rowCountCache(runner: DBRunner)(builder: Caffeine[AnyRef, AnyRef])(
+      implicit ec: ExecutionContext): CaffeineAsyncCache[Query[_, _, Seq], Int] =
+    CaffeineAsyncCache {
+      builder
+        .asInstanceOf[Caffeine[Query[_, _, Seq], Int]]
+        .buildAsync[Query[_, _, Seq], Int] {
+          new AsyncCacheLoader[Query[_, _, Seq], Int] {
+            override def asyncLoad(table: Query[_, _, Seq],
+                                   executor: Executor): CompletableFuture[Int] =
+              runner.run(table.length.result).asJava.toCompletableFuture
+          }
+        }
+    }
 }
 
 /** A wrapper around [[com.github.benmanes.caffeine.cache.AsyncLoadingCache]] to convert Java types to Scala. */
