+++ app/src/main/scala/org/alephium/explorer/persistence/dao/BlockDao.scala
@@ -16,8 +16,6 @@
 
 package org.alephium.explorer.persistence.dao
 
-import java.util.concurrent.{CompletableFuture, Executor}
-
 import scala.concurrent.{ExecutionContext, Future}
 import scala.jdk.CollectionConverters._
 import scala.jdk.FutureConverters._
@@ -134,39 +132,11 @@ object BlockDao {
       .expireAfterWrite(5, java.util.concurrent.TimeUnit.SECONDS)
       .buildAsync(blockTimeAsyncLoader)
 
-    /** Caches the number of rows return by a query.
-      *
-      * @note Slick's [[Query]] type does not implement equals or hashCode so
-      *       this cache will not work as expected on dynamically generated queries.
-      *       The queries to cache should be static i.e. should have the same memory
-      *       address.
-      *
-      *        {{{
-      *          //OK
-      *          cacheRowCount.get(mainChainQuery)
-      *          cacheRowCount.get(mainChainQuery)
-      *
-      *          //OK
-      *          cacheRowCount.get(BlockHeaderSchema.table)
-      *          cacheRowCount.get(BlockHeaderSchema.table)
-      *
-      *          //NOT OK
-      *          cacheRowCount.get(BlockHeaderSchema.table.filter(_.mainChain))
-      *          cacheRowCount.get(BlockHeaderSchema.table.filter(_.mainChain))
-      *        }}}
-      * */
     private val cacheRowCount: CaffeineAsyncCache[Query[_, _, Seq], Int] =
-      CaffeineAsyncCache {
+      CaffeineAsyncCache.rowCountCache(this) {
         Caffeine
           .newBuilder()
           .refreshAfterWrite(2, java.util.concurrent.TimeUnit.MINUTES)
-          .buildAsync[Query[_, _, Seq], Int] {
-            new AsyncCacheLoader[Query[_, _, Seq], Int] {
-              override def asyncLoad(table: Query[_, _, Seq],
-                                     executor: Executor): CompletableFuture[Int] =
-                run(table.length.result).asJava.toCompletableFuture
-            }
-          }
       }
 
     def getRowCountFromCacheIfPresent[E, U](query: Query[E, U, Seq]): Option[Future[Int]] =
